// src/generated/client/models/Resource.ts
export interface Resource {
    /**
     * Resource id.
     * @type {string}
     * @memberof Resource
     */
    readonly id?: string;
    /**
     * Resource parent id.
     * @type {string}
     * @memberof Resource
     */
    parentId?: string;
    /**
     * 
     * @type {number}
     * @memberof Resource
     */
    orderNumber?: number;
    /**
     * 
     * @type {ResourceType}
     * @memberof Resource
     */
    type: ResourceType;
    /**
     * 
     * @type {Array<KeyValueProperty>}
     * @memberof Resource
     */
    styles?: Array<KeyValueProperty>;
    /**
     * 
     * @type {Array<KeyValueProperty>}
     * @memberof Resource
     */
    properties?: Array<KeyValueProperty>;
    /**
     * Resource display name.
     * @type {string}
     * @memberof Resource
     */
    name: string;
    /**
     * Resource data, either URL on images, videos, etc and text data on text resources
     * @type {string}
     * @memberof Resource
     */
    data?: string;
    /**
     * Resource slug name.
     * @type {string}
     * @memberof Resource
     */
    slug: string;
    /**
     * Creation time
     * @type {Date}
     * @memberof Resource
     */
    readonly createdAt?: Date;
    /**
     * Last modification time
     * @type {Date}
     * @memberof Resource
     */
    readonly modifiedAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof Resource
     */
    creatorId?: string;
    /**
     * 
     * @type {string}
     * @memberof Resource
     */
    lastModifierId?: string;
}